Add files via upload
The UCI Zoo dataset is a popular benchmark dataset commonly used for classification tasks. It contains information about 101 different animals, categorized into seven classes. The dataset provides a variety of features that describe the animals, such as their hair, feathers, eggs, milk, airborne capabilities, aquatic nature, predator status, and more.

Here are some details about the UCI Zoo dataset:

Number of Instances: There are a total of 101 instances in the dataset, each representing a different animal.

Number of Features: The dataset consists of 18 features (or attributes) that describe each animal. These features include:

Animal name: The name of the animal (nominal).
Hair: Whether the animal has hair or not (binary: "0" or "1").
Feathers: Whether the animal has feathers or not (binary: "0" or "1").
Eggs: Whether the animal lays eggs or not (binary: "0" or "1").
Milk: Whether the animal produces milk or not (binary: "0" or "1").
Airborne: Whether the animal can fly or not (binary: "0" or "1").
Aquatic: Whether the animal lives in water or not (binary: "0" or "1").
Predator: Whether the animal is a predator or not (binary: "0" or "1").
Toothed: Whether the animal has teeth or not (binary: "0" or "1").
Backbone: Whether the animal has a backbone or not (binary: "0" or "1").
Breathes: Whether the animal breathes air or not (binary: "0" or "1").
Venomous: Whether the animal is venomous or not (binary: "0" or "1").
Fins: Whether the animal has fins or not (binary: "0" or "1").
Legs: The number of legs the animal has (numeric: 0, 2, 4, 5, 6, 8).
Tail: Whether the animal has a tail or not (binary: "0" or "1").
Domestic: Whether the animal is domestic or not (binary: "0" or "1").
Catsize: Whether the animal is of catsize or not (binary: "0" or "1").
Classes: The animals in the dataset are classified into seven different classes:

Mammal
Bird
Reptile
Fish
Amphibian
Bug
Invertebrate
The UCI Zoo dataset is commonly used for training and evaluating classification algorithms. Researchers and machine learning practitioners use this dataset to develop models that can accurately predict the class of an animal based on its features. It serves as a good starting point for learning and practicing classification techniques and algorithms.